Source/backend/requirements.txt
rq==1.5.2
pymongo==3.10.1
pulsar-client==2.6.1
flask==1.1.2
flask_cors==3.0.8
gunicorn==20.0.4